Unipile chatbot/requirements.txt
openai
requests
python-dotenv
rich
pydantic
